Move helm.proxy.tokenizers to helm.tokenizers
