+++ MANIFEST.in
@@ -1,5 +1,5 @@
 recursive-include src/helm/ py.typed
-recursive-include src/helm/proxy/clients/ *.sp
+recursive-include src/helm/tokenizers/ *.sp
 recursive-include src/helm/benchmark/ *.json
 recursive-include src/helm/benchmark/static/ *.css *.html *.js *.png *.yaml
 recursive-include src/helm/config/ *.yaml
