+++ src/helm/proxy/token_counters/auto_token_counter.py
@@ -2,7 +2,7 @@ from typing import List
 from helm.benchmark.model_deployment_registry import ModelDeployment, get_model_deployment
 
 from helm.common.request import Request, Sequence
-from helm.proxy.tokenizers.auto_tokenizer import AutoTokenizer
+from helm.tokenizers.auto_tokenizer import AutoTokenizer
 from helm.common.tokenization_request import TokenizationRequest, TokenizationRequestResult
 from .token_counter import TokenCounter
